Add 116, Remove pointless "#include"
Added solution to problem 116, Medium
Populating Next Right Pointers in Each Node
Populate each next pointer to point to its next right node. If there is no next right node, the next pointer should be set to NULL.

Also removed the include statements as this is not meant to be a stand alone program, rather contains the solutions to be placed into leetcode.